Stop installing Python2 due to EOL
New Macs still have Python 2.7 as the system python.

If we ever run into the need to have a specific v2 version, can set it
up manually. It's unlikely at this point.

Python 2.7 reached the end of its life on January 1st, 2020.
Pip 21.0 will drop support for Python 2.7 in January 2021.