Verify GH host key during SSH key setup phase
So the script stays idempotent in the sense that we don't try to reauthenticate if we already did on a previous script run.

Note we fallback with "true" because ssh will return the code from the remote command, and GitHub returns 1 because it
does not provide SSH shell access. We don't want that exit code to abort our script.