Rewrite asdf script with Claude Code
The old script needed to be replaced since it no
longer worked with asdf v0.16 and higher. They
rewrote asdf from a series of bash scripts to
golang and changed the way it's installed.

Good opportunity to test Claude Code which is in
research preview. I changed a few things from the
generated code but overall a good experience.